+++ src/main/java/com/arcanc/pulselib/content/model/animation/PKeyFrameChannel.java
@@ -18,17 +18,59 @@ public sealed interface PKeyFrameChannel<T> permits PKeyFrameChannel.PositionKey
 													PKeyFrameChannel.ScaleKeyFrame
 {
 	float time();
-	T value();
 	
-	record PositionKeyFrame(float time, Vector3f value) implements PKeyFrameChannel<Vector3f>
+	T value(Object data);
+
+	default T value()
 	{
+		return value(null);
+	}
+
+	record PositionKeyFrame(float time, ValueProvider<Vector3f> valueProvider) implements PKeyFrameChannel<Vector3f>
+	{
+		public PositionKeyFrame(float time, Vector3f value)
+		{
+			this(time, data -> value);
+		}
+
+		@Override
+		public Vector3f value(Object data)
+		{
+			return this.valueProvider.calculate(data);
+		}
+	}
+
+	record RotationKeyFrame(float time, ValueProvider<Quaternionf> valueProvider) implements PKeyFrameChannel<Quaternionf>
+	{
+		public RotationKeyFrame(float time, Quaternionf value)
+		{
+			this(time, data -> value);
+		}
+
+		@Override
+		public Quaternionf value(Object data)
+		{
+			return this.valueProvider.calculate(data);
+		}
+	}
+
+	record ScaleKeyFrame(float time, ValueProvider<Vector3f> valueProvider) implements PKeyFrameChannel<Vector3f>
+	{
+		public ScaleKeyFrame(float time, Vector3f value)
+		{
+			this(time, data -> value);
+		}
+
+		@Override
+		public Vector3f value(Object data)
+		{
+			return this.valueProvider.calculate(data);
+		}
 	}
 	
-	record RotationKeyFrame(float time, Quaternionf value) implements PKeyFrameChannel<Quaternionf>
+	@FunctionalInterface
+	interface ValueProvider<T>
 	{
-	}
-	
-	record ScaleKeyFrame(float time, Vector3f value) implements PKeyFrameChannel<Vector3f>
-	{
+		T calculate(Object data);
 	}
 }