+++ changelog.md
@@ -1,2 +1,1 @@
-Now dynamic color can be applied to anything
-Removed duplicated class. Simple clean up and a bunch of renamed
+* Added Molang animation support